// WebMap.Web/src/net/models/message-type.ts
// @ts-nocheck
// automatically generated by the FlatBuffers compiler, do not modify

/* eslint-disable @typescript-eslint/no-unused-vars, @typescript-eslint/no-explicit-any, @typescript-eslint/no-non-null-assertion */

export enum MessageType {
  MapData = 0,
  PlayerLocation = 1,
  PlayerPosition = 2,
  SelectedLocation = 3,
  DiscoveryMask = 4,
  QuestMarkers = 5,
  UnacceptedQuestMarkers = 6,
  QuestLinkMarkers = 7,
  LevequestMarkers = 8,
  ActiveLevequestMarker = 9,
  GuildLeveAssignmentMarkers = 10,
  GuildOrderGuideMarkers = 11,
  HousingMarkers = 12,
  TripleTriadMarkers = 13,
  CustomTalkMarkers = 14,
  GemstoneTraderMarkers = 15,
  FlagMarkers = 16
}